I just realized that the simplified gameplay doesn't fully match the regular one.
If simplified is assuming that you want to move the stack first (and
only want to select a card from the stack to move, second) and I pick
a number higher than the number of face-up cards in the stack, the
game should assume I was trying to pick a stack for which all of the
cards cannot be moved.

This is, also, useful because I use it, all the time, to quickly
cancel a stack selection, if I didn't mean to try and move the stack I
just selected (rather than use the more involved =C-g= key-binding).

In any case, I think that'd be more expected and not remove you from
the gameplay, as much, so let's fix that.

I wish I didn't have to put what's essentially ~klondike--card-move~
logic in ~klondike--stack-pick~ but I don't think it makes /any/ sense
to repeat the logic of checking how many face-up cards are in the
stack outside of ~klondike--stack-pick~.

It's sort of fragmentory but, eh, that's where imperative coding
leads…

And we need to be imperative (at least, without getting exceedingly
more creative) and have the global variables because we need to keep
track of state while we're waiting for input. Given Emacs's structure,
we enter a mode and then do nothing while we wait for input from the
user; so we need to keep track of the state of the world, somehow,
when we, essentially, enter idle.

At least, my documentation has been fairly decent to help explain how
everything's connected (and why).